distillation/requirements.txt
numpy==1.21.6
Pillow==9.3.0
scikit_learn==1.1.3
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
vit_pytorch==0.38.1
